backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
tortoise-orm==0.20.0
aiomysql==0.2.0
aerich==0.7.2

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Excel Processing
openpyxl==3.1.2
pandas==2.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
